Ruby style: use full words over letters
The particular bit of code that prompted this came from BulkAssignment.rb

```ruby
@assignee.users.each do |user|
  assignables.map do |a|
    assignment = a.assign_to(user)
```

Here, we’re dealing with an assignee, assignables, and assignments, so it’s particularly unclear what |a| is referring to. But in general, I think this is a good practice. I’m surprised neither Bbatsov’s nor Github’s styleguides mention this, as I swear I’ve read it elsewhere.